Optimize StreamBuffer when empty
To avoid unnecessary copies, StreamBuffer writes at most until the "end"
of the circular buffer, which is subobtimal (it could have written more
data if they have been contiguous).

In order to minimize the occurrence of this event, reset the head and
tail to 0 when the buffer is empty (no copy is involved).

This is especially useful when the StreamBuffer is used to read/write
one packet at a time, so the "end" of the buffer is guaranteed to never
be reached.