Prepare deferred packet sending to client
A packet built from the data received from the network TCP socket may
not be sent to the client for several reasons:
 - when the client buffer is full, sendToClient(…) returns false, and
   the packet is not sent;
 - when the packet sequence number is outside the client window, we must
   keep it until the window advances enough.

Those packets may not be dropped, because they are received from the
network TCP socket: we have the responsability to transmit it,
otherwise the whole connection would be broken.

Therefore, when a packey may not be sent, keep it, and disable our
interest in reading the network TCP socket. When the client is ready to
receive more packets, it will call processPacketForClient() (not
implemented yet).

We must take care of updating the acknowledgement number (we may have
received packets since we built our pending IPv4Packet).