Implement packet source consuming
Connections may use sendToClient(packet) to send a packet to the client.
However, this may fail (and return false), e.g. when the client buffer
is full.

This is ok for our crafted empty packet (ACK, RST, SYN-ACK…), because
their loss is gracefully handled.

However, we cannot lose packets built from the network socket streams,
because there is no way to recover. We also want to avoid new
allocations for keeping them, because they would be unlimited.

Therefore, instead of sending a packet, a TCPConnection (the packet
source) requests the client to pull from it, by calling consume(this).
The reading interest from the network socket will be disabled until the
client succeeds.

As soon as the client buffer get more free space, the client takes the
responsability to make another attempts to pull from the pending packet
sources.