Provide access to host "localhost" via 10.0.2.2
From the device, local addresses like 127.0.0.1 refer to the device
itself. Non-local addresses are forwarded through the tunnel to the
host.

Therefore, there was no way to refer to the host "localhost". In
particular, on Ubuntu, the default DNS server points to 127.0.1.1.

To overcome this problem, map 10.0.2.2 to localhost like the Android
emulator does. Therefore, every TCP or UDP connection having 10.0.2.2 as
destination will open a socket to InetAddress.getLoopbackAddress() from
the host.