Wake up blocking read() using a single UDP packet
In order to wake up a blocking read() on VPN, we triggered a name
resolution request, that in turn sent a UDP packet.

However, a name resolution may send several UDP packets successively,
because it retries if it gets no response. In that case, some UDP
packets may actually be sent over the network, which is unnecessary.

Instead, send a single meaningless UDP packet, just for waking up the
blocking read().