Fix broken state after RST due to duplicate SYN
When we received a duplicate SYN (in a state different from SYN_SENT),
the connection was reset (the state was set to null and a RST packet was
sent). However, contrary to what happens after the other calls to
resetConnection(), destroy() was not called.

Thus, the TCPConnection state was reset, but the socket channel was not,
so it might be left connected. As a consequence, the next call to
handleFirstPacket() will set the state to SYN_SENT, leading OP_CONNECT
to be set as interesting op in updateInterests().

Registering OP_CONNECT to a socket already connected is incorrect, it
will wake up selector.select() at every call, leading to 100% CPU usage.

Therefore, call destroy() directly in resetConnection(), since not
calling it leaves the TCPConnection in inconsistent state.

But then, duplicate SYNs (which are not so uncommon) always reset the
connection, making the relaying unusable. To avoid this issue, reset
connection only for "unexpected" duplicate SYNs, i.e. a SYN having a
sequence number different from the one received during the
initialization of this connection.