Restrict GnirehtetControlService to privilegied
We want to avoid the VPN to be started or stopped by anyone, so it must
require a permission.

Since intercepting the whole network traffic is critical, we want a
system permission.

In addition, we want to call it from shell, so we need an existing
system permission granted to shell:
<https://github.com/android/platform_frameworks_base/blob/master/packages/Shell/AndroidManifest.xml>

From the small permissions set matching these conditions, I chose
android.permission.WRITE_SECURE_SETTINGS.

Note: The following network-related system permissions granted to shell
may sound more appropriate:
 - android.permission.MANAGE_NETWORK_POLICY
 - android.permission.CONNECTIVITY_INTERNAL

Unfortunately, they were added only recently, so they are not granted to
shell for older versions:
<https://github.com/android/platform_frameworks_base/commit/6a05eee2bdbabb844f1d63741a0680477ec63546>