Fix serial management in "gnirehtet rt"
The command "gnirehtet rt" first checks whether the app is installed on
the device, and installs it if necessary.

However, it didn't forward the serial parameter, so calling "gnirehtet
rt 12345678" failed to install the application on the selected device
when several were plugged.

Therefore, forward the serial to apk_require_installed().