Make gnirehtet callable from everywhere
The gnirehtet script assumed that it was called from its own directory.

To overcome this limitation, detect and move to the physical directory
of the script. This requires to resolve symlinks first (in case the
script is called through a symlink).

Unfortunately, "realpath" is not always installed by default, and the
"readlink" behavior is different on Mac OS. Therefore, implement a best
effort solution.

See <http://stackoverflow.com/a/17744637/1987178>.

Related to <https://github.com/Genymobile/gnirehtet/issues/6>.