Support API >= 14
The minSdkVersion was set to 21 due to the call to
VpnService.Builder.setBlocking(true).

On API < 21, we can call the internal method
libcore.io.IoUtils.setBlocking() using reflection:
<https://android.googlesource.com/platform/libcore/+/30c669166d86d0bd133edfb67909665fb41d29b6/luni/src/main/java/libcore/io/IoUtils.java#89>

Therefore, use it as a fallback and set the minSdkVersion to 14, in
order to support devices from Android 4.0.

Related to <https://github.com/Genymobile/gnirehtet/issues/2>.