Release the mutex for waiting
During the delay between connection attempts, the RelayTunnelProvider
must release the mutex (by calling wait() instead of sleep()), so that
invalidateTunnel() may be called at any time.

Indeed, there was a lock dependency that prevented the tunnel to be
closed immediately. On gnirehtet stop, the following cascade of calls
occurred:
 -> GnirehtetService.close()
   -> Forwarder.stop()
     -> PersistentRelayTunnel.close()
       -> RelayTunnelProvider.invalidateTunnel() // synchronized
          // required the mutex that may be held by getCurrentTunnel()
          // during the whole execution of Thread.sleep()
     -> Future.cancel(true) // interrupts Thread.sleep()

Thread.sleep() was intended to be interrupted by a call to
Future.cancel(true). However, this call occurred only after
PersistentRelayTunnel.close() completed… that is, after the end of
Thread.sleep(), making the whole interruption useless.