Fix tunnel reconnection race condition
On the device-side, sending to and receiving from the tunnel (the
connection to the relay server) are processed from two different
threads.

When an error occurs, both threads invalidate the tunnel, so that the
next call to getCurrentTunnel() will return a new tunnel. Typically, the
first one actually invalidates the tunnel, then the second one detects
that the tunnel has already been invalidated, and does nothing.

However, there is a possibility that one thread invalidates the tunnel,
then calls getCurrentTunnel() before the other thread calls
invalidateTunnel(). In that case, the new tunnel would be mistakenly
invalidated.

To prevent this race condition, pass the current tunnel as parameter of
invalidateTunnel().