Early detect invalid connection to relay server
The relay server is accessible through an "adb reverse" port
redirection.

If the port redirection is enabled but the relay server is not started,
then the call to channel.connect() will succeed, but the first read()
will return -1.

As a consequence, the connection state of the relay server would be
invalid temporarily (we would switch to CONNECTED state then switch back
to DISCONNECTED).

To avoid this problem, we must actually try to read from the server, so
that an error occurs immediately if the relay server is not accessible.

To do so, we set the lowest timeout possible (1) that is not "infinity"
(0), try to read, then reset the timeout option. Since the client is
always the first peer to send data, there is never anything to read.