Release mutex during blocking calls
RelayTunnel.open() both opened a channel and connected the socket to the
relay server (or, more precisely, to the tunnel created by "adb reverse"
that enabled port forwarding to the relay server).

However, connecting the socket is blocking, and was called from
RelayTunnelProvider.getCurrentTunnel() with the mutex locked. As a
consequence, it was not possible to call invalidateTunnel() (and, a
fortiori, to close the VPN) until connect() completed.

To avoid this problem, split the channel opening and the socket
connection, so that connect() may be called with the mutex (the one used
by invalidateTunnel()) released.

However, we must still make sure that no more than one thread executes
getCurrentTunnel() at the same time. For this purpose, use a second
mutex (getCurrentTunnelLock()).