Improve "closing" naming consistency
A Client and a Route executed a callback (RemoveHandler) provided by
their "owner" when they have to be removed.

Instead, consider this callback as a listener, so rename it
"CloseListener". The fact that their owners "remove" them from some
lists is irrelevant for Client and Route: they have only the
responsibility to notify them when they must not be used anymore.

For consistency, also rename "discard" and "destroy" to "close".