Type-erase CloseListener in Client::new()
Client::new() was a generic function to accept statically any
CloseListener<Client>. That way, the caller may directly pass a closure.

However, this implies that there will be one version of this function
for every CloseListener type. To prevent this, type-erase the listener
instead.

Note that currently, there is exactly one call to Client::new(), so
there is only one version. But the intent is to get one version whatever
the number of different CloseListener types.