Dynamically dispatch connections
Make Connection a trait.

Connection was implemented to statically dispatch over TCPConnection and
UDPConnection.

However, these instance need to be shared between the Route and the
selector, so they are embedded into an Rc<RefCell<…>>.

Mixing static dispatch with indirection adds complexity and is not
elegant. Therefore, switch to dynamic dispatch.