[DO_NOT_MERGE] Never keep UDP packet for client
As it is currently implemented, if the client buffer is full, the UDP
connection keep the packet and disable reading, but contrary to
TCPConnection, there is no mechanism to retry once the client buffer is
not full anymore.

This commit will help to confirm issue #29, but is not intended to be
merged as is.