Drop UDP packets when client buffer is full
For TCP connections, we may not lose data we receive from the network.
When the client buffer is full, we cannot transmit it to the client, so
we "pause" this specific connection: we keep the packet and disable read
operation for the channel. When the client is ready to receive more
data, it request to "paused" connections (through PacketSource).

UDPConnection also kept the packet and disabled read operation for the
channel. This was wrong, because it didn't use the PacketSource
mechanism, so there was no way to "unpause" such a connection.

Since we may lose UDP datagrams, just drop a packet when the client
buffer is full instead.