Introduce ClientChannel
A connection (e.g. TCPConnection) has a Rc/RefCell to its client.

When a Client sends a TCP packet, the TCPConnection may immediately
reply (typically with an empty packet with flags set).

However, the TCPConnection may not borrow its client RefCell, since the
client is already borrowed by the caller.

As a consequence, the caller must give the possibility to reply a packet
immediately.

Therefore, provide ClientChannel, a lifetime-bounded view of the client
StreamBuffer, so that the connection may send packets to the borrowed
client.