Move poll interruption handling to Main
On Ctrl+C, poll() will return an error with ErrorKind::Interrupted.
Instead of handling it locally, throw it up until main, which catch the
error.

In case of interruption, wait a bit before exiting, so that the handler
execution is complete. This avoids to print logs to the terminal after
the prompt is available.