Accept empty interest in mio
Currently, the mio library does not accept registering a handle with
empty interest. As a workaround, we provided a wrapper that exposed an
API that did, calling deregister/reregister internally as necessary.

However, this workaround does not work on Windows, because
deregister/reregister is not implemented:
<https://github.com/carllerche/mio/issues/633>

Therefore, make mio accept empty interest directly:
<https://github.com/carllerche/mio/pull/640>

This is implemented in our own fork, unless this change is accepted
upstream or we manage to find a solution that works with the upstream
version.