Retry on Interrupted error
A blocking call may return an Err having "Interrupted" as kind(),
typically when the user press Ctrl+C in "gnirehtet rt". The error was
thrown and the relay server was interrupted.

However, an "Interrupted" error may occur in other cases. Concretely,
using a cpuprofiler causes poll() to generate such errors sometimes.

Therefore, retry on "Interrupted" when no interruption was requested by
the user.