Optimize Rust TCP checksum
Checksums computation is the most CPU-intensive task in gnirehtet, so it
needs to be optimized.

Firstly, do not use a Cursor (a non-zero-cost abstraction).

Secondly, for summing 16-bit values, sum high-order and low-order bytes
separately. This allow to do the shift operation only once (instead of
once per value).

The new implementations runs ~4.4× faster (i.e. it takes ~23% of the
time required by the old implementation).