Optimize more Rust TCP checksum with unsafe
For summing array items, use unchecked accesses to the array, in order
to avoid bounds checking for every value.

This runs ~4.1× faster than with bounds checking (i.e. it takes ~24% of
the time necessary when bounds checking is enabled).

Overall (including the previous optimizations), the resulting
implementation is ~18× faster (it runs in 5.5% of the time required by
the old implementation).

Since this is the most CPU-intensive task, gnirehtet uses significantly
less CPU than before.