Fix TCP checksum computation
The last optimization broke TCP checksums, in a way that was not
captured by unit tests.

In an unsafe block, we computed an offset of (payload_length - 1).
However, payload_length is u16 (unsigned), so the result was wrong when
payload_length is 0.

Fix the offset, and add a unit test for checksuming empty payloads.