Simplify Ctrl+C termination
The interruption handler attempted to make the relay() call complete, so
that gnirehtet exited normally. This added a lot of complexity, and was
not free of race conditions: in the ctrlc crate implementation, the user
handler is not called from the OS handler, so setting a flag in the user
handler was inherently racy.

Instead, just call exit() from the interrupt handler.

This is far more simple and works perfectly.