Use headers data directly when possible
To route a packet to an existing connection, the Router only needs the
headers data, not the whole headers (which include a reference to the
raw array accessible through a "borrowing view").

The whole headers are necessary only when it creates a new connection.

Therefore, make Router use references to headers data directly until it
has to create a new Connection.