Improve error handling
In CommandExecutionError, "command" was ambiguous, it referred both to:
 - a gnirehtet command (start, stop, run, relay…),
 - a process command (the execution of "adb something").

CommandExecutionError was either an Io error or a Status error. When it
is a process Io error, we want to report the failing command.

Therefore, refactor the CommandExecution enum with 3 values:
 - ProcessIo,      // io::Error due to process command execution
 - ProcessStatus,  // a process command returned with non-zero value
 - Io              // any other io::Error

Thus, ProcessIo can provide the failing command, to be reported to the
user.

Also, move all the execution error classes to a separate module.