Fix u16 overflow in IPv4PacketBuffer
When the buffer is full, self.head equals 0x10000, so converting its
value to u16 results in 0.

As a consequence, length <= self.head was mistakenly considered false,
so the packet was never consumed.

As a consequence, the next call to read_from() requested a read with
a buffer of 0 bytes, so 0 bytes were read… which was interpreted as EOF,
causing client disconnection.

See <https://github.com/Genymobile/gnirehtet/issues/43>.