Terminate TCP connections gracefully
When the network socket is closed or when the device sends a FIN,
gracefully terminate the connection by following the RFC793 instead of
closing without notifying the device.

This avoids to receive more unwanted packets from the device, leading to
connections open and close on first packet.

See schema: <https://tools.ietf.org/html/rfc793#page-23>