Ack only once data is written to network socket
The TCP connection acked packets received from the device immediately
after it stored them into its buffers, but before they were written to
the network socket.

This made the client wrongly assume that the connection was super-fast,
so the upload bandwidth was totally uncorrelated to the network socket
bandwidth.

As a consequence, the client filled the buffers up to saturation,
leading to packet drops, that slowed down uploads dramatically.

To make the client adapt the bittrate according to the associated
network socket, do not ack the data immediately, but only once it is
actually written to the network socket.

See issues:
 - <https://github.com/Genymobile/gnirehtet/issues/84>
 - <https://github.com/Genymobile/gnirehtet/issues/79>