Ignore double-removal of a connection
In a single poll() result, the client may be ready to receive a packet
from the device, that will be sent to a specific network connection, and
in the same time that connection may be ready to send a packet to the
client.

There is a possibility that both events will lead to close the
connection, and remove it from the router. In that case, the second
removal did panic.

Instead, remove it only if it does not exist.

See:
 - <https://github.com/Genymobile/gnirehtet/issues/61>
 - <https://github.com/Genymobile/gnirehtet/issues/85>